# openapi.yaml
openapi: 3.1.0
info:
  title: Social Neuron REST API
  version: 1.5.2
  description: |
    REST API for Social Neuron — AI-powered social media content creation, scheduling, and analytics.

    ## Authentication
    All endpoints require an API key via Bearer token:
    ```
    Authorization: Bearer snk_live_...
    ```
    Get your API key at [socialneuron.com/settings/developer](https://socialneuron.com/settings/developer).

    ## Scopes
    API keys have scopes that control access:
    - `mcp:full` — All permissions
    - `mcp:read` — Read analytics, brand, credits
    - `mcp:write` — Generate content, save plans
    - `mcp:distribute` — Schedule posts
    - `mcp:analytics` — Refresh analytics
    - `mcp:comments` — Manage comments
    - `mcp:autopilot` — Automated scheduling (Pro+)

    ## Async Operations
    Video/image generation returns `202 Accepted` with a `Location` header pointing to the job status endpoint.
    Poll `GET /v1/jobs/{id}` with the `Retry-After` interval until status is `completed` or `failed`.

    ## Rate Limits
    - Read operations: 60 requests/minute
    - Write/posting operations: 30 requests/minute
    - IP-level: 60 requests/minute (pre-auth)

    Rate-limited responses include `Retry-After` header.
  contact:
    name: Social Neuron
    url: https://socialneuron.com
    email: developers@socialneuron.com
  license:
    name: MIT
servers:
  - url: https://mcp.socialneuron.com/v1
    description: Production
  - url: http://localhost:8080/v1
...
security:
  - apiKey: []
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: "API key with snk_live_ prefix"
  schemas:
    Error:
      type: object
      required: [error, error_description, status]
      properties:
        error:
          type: string
          example: validation_error
        error_description:
          type: string
          example: "prompt is required"
        status:
          type: integer
          example: 400
        retry_after:
          type: integer
          description: Seconds to wait before retrying (rate limit responses)
...
    Platform:
      type: string
      enum: [youtube, tiktok, instagram, twitter, linkedin, facebook, threads, bluesky]
    VideoModel:
      type: string
      enum: [veo3-fast, veo3-quality, runway-aleph, sora2, sora2-pro, kling, luma, midjourney-video]
    ImageModel:
      type: string
      enum: [midjourney, nano-banana, nano-banana-pro, ideogram, flux-pro, flux-max, gpt4o-image, imagen4, imagen4-fast, seedream]
    AspectRatio:
      type: string
      enum: ["16:9", "9:16", "1:1", "4:3", "3:4"]
...
paths:
...
  /content/generate:
    post:
      summary: Generate text content
      tags: [Content]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt:
                  type: string
                  description: Content generation prompt
                platform:
                  $ref: "#/components/schemas/Platform"
                content_type:
                  type: string
                  enum: [script, caption, blog, hook, generation]
                tone:
                  type: string
                brand_voice:
                  type: string
                project_id:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Generated content
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /content/video:
    post:
      summary: Generate video (async)
      tags: [Content]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt:
                  type: string
                model:
                  $ref: "#/components/schemas/VideoModel"
                aspect_ratio:
                  $ref: "#/components/schemas/AspectRatio"
                duration:
                  type: integer
                  minimum: 1
                  maximum: 60
                reference_image_url:
                  type: string
                  format: uri
      responses:
        "202":
          description: Job accepted
          headers:
            Location:
              schema:
                type: string
              description: URL to poll for job status
            Retry-After:
              schema:
                type: integer
              description: Seconds to wait before polling
  /content/image:
    post:
      summary: Generate image (async)
      tags: [Content]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt:
                  type: string
                model:
                  $ref: "#/components/schemas/ImageModel"
                aspect_ratio:
                  $ref: "#/components/schemas/AspectRatio"
                style:
                  type: string
                negative_prompt:
                  type: string
      responses:
        "202":
          description: Job accepted
  /content/carousel:
    post:
      summary: Generate carousel content
      tags: [Content]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [topic]
              properties:
                topic:
                  type: string
                platform:
                  $ref: "#/components/schemas/Platform"
                slides:
                  type: integer
                  default: 5
                brand_voice:
                  type: string
                project_id:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Carousel content
  /content/voiceover:
    post:
      summary: Generate AI voiceover
      tags: [Content]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                voice:
                  type: string
                language:
                  type: string
                  default: en
      responses:
        "200":
          description: Voiceover audio
  /content/adapt:
    post:
      summary: Adapt content for platforms
      tags: [Content]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content, target_platforms]
              properties:
                content:
                  type: string
                source_platform:
                  $ref: "#/components/schemas/Platform"
                target_platforms:
                  type: array
                  items:
                    $ref: "#/components/schemas/Platform"
                brand_voice:
                  type: string
      responses:
        "200":
          description: Adapted content
  /content/trends:
    get:
      summary: Fetch trending topics
      tags: [Content]
      parameters:
        - name: source
          in: query
          schema:
            type: string
            enum: [youtube, google_trends, rss, url]
            default: youtube
        - name: category
          in: query
          schema:
            type: string
            default: all
        - name: region
          in: query
          schema:
            type: string
            default: US
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Trending topics

  # ── Posts ─────────────────────────────────────────────────────────
...
  /plans:
    post:
      summary: Generate content plan
      tags: [Plans]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [topic, platforms]
              properties:
                topic:
                  type: string
                platforms:
                  type: array
                  items:
                    $ref: "#/components/schemas/Platform"
                days:
                  type: integer
                  default: 7
                brand_voice:
                  type: string
                source_url:
                  type: string
                  format: uri
                project_id:
                  type: string
                  format: uuid
      responses:
        "201":
          description: Content plan created
    get:
      summary: List content plans
      tags: [Plans]
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: project_id
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of plans
  /plans/approvals:
    get:
      summary: List pending plan approvals
      tags: [Plans]
      parameters:
        - name: plan_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Plan approvals
  /plans/{id}:
    get:
      summary: Get a content plan
      tags: [Plans]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Content plan
        "404":
          description: Plan not found
    put:
      summary: Update a content plan
      tags: [Plans]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                posts:
                  type: array
                topic:
                  type: string
                status:
                  type: string
      responses:
        "200":
          description: Plan updated
  /plans/{id}/schedule:
    post:
      summary: Schedule plan posts
      tags: [Plans]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                auto_slot:
                  type: boolean
                  default: true
                batch_size:
                  type: integer
                  default: 5
                dry_run:
                  type: boolean
                  default: false
      responses:
        "202":
          description: Scheduling initiated
  /plans/{id}/approve:
    post:
      summary: Approve/reject plan
      tags: [Plans]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [approve, reject, needs_edit]
                  default: approve
                post_ids:
                  type: array
                  items:
                    type: string
                feedback:
                  type: string
      responses:
        "200":
          description: Approval response

  # ── Comments ─────────────────────────────────────────────────────
...
  /jobs/{id}:
    get:
      summary: Poll async job status
      tags: [Jobs]
      description: |
        Poll the status of an async job (video/image generation).
        Returns `Retry-After` header while job is in progress.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job status with result if completed

  # ── Account ──────────────────────────────────────────────────────
...